my_flask_project-2/requirements.txt
absl-py==2.1.0
altgraph==0.17.4
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
astunparse==1.6.3
attrs==23.2.0
backcall==0.2.0
bcrypt==4.1.3
beautifulsoup4==4.12.3
bleach==6.0.0
cachetools==5.3.3
celluloid==0.2.0
cffi==1.15.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.1.4
cryptography==42.0.7
cycler==0.11.0
debugpy==1.7.0
decorator==4.4.2
defusedxml==0.7.1
entrypoints==0.4
exceptiongroup==1.2.1
fastjsonschema==2.19.1
ffmpeg==1.4
Flask==2.2.5
Flask-SQLAlchemy==3.0.5
flatbuffers==24.3.25
fonttools==4.38.0
gast==0.5.4
google-auth==2.29.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.62.2
h5py==3.8.0
idna==3.7
imageio==2.31.2
imageio-ffmpeg==0.5.1
importlib-metadata==6.7.0
importlib-resources==5.12.0
ipykernel==6.16.2
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==8.1.3
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.4
joblib==1.3.2
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-server==1.24.0
jupyter_client==7.4.9
jupyter_core==4.12.0
jupyterlab-pygments==0.2.2
jupyterlab_widgets==3.0.11
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.4.4
MarkupSafe==2.1.5
matplotlib==3.5.3
matplotlib-inline==0.1.6
mistune==3.0.2
moviepy==1.0.3
mpmath==1.3.0
natsort==8.4.0
nbclassic==1.1.0
nbclient==0.7.4
nbconvert==7.6.0
nbformat==5.8.0
nest-asyncio==1.6.0
notebook==6.5.7
notebook_shim==0.2.4
numpy==1.21.6
oauthlib==3.2.2
opencv-python==4.9.0.80
opt-einsum==3.3.0
packaging==24.0
pandas==1.3.5
pandocfilters==1.5.1
paramiko==3.4.0
parso==0.8.4
pefile==2024.8.26
pickleshare==0.7.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
proglog==0.1.10
prometheus-client==0.17.1
prompt_toolkit==3.0.45
protobuf==3.20.3
psutil==5.9.8
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.17.2
pyinstaller==5.13.2
pyinstaller-hooks-contrib==2024.7
PyMySQL==1.1.1
PyNaCl==1.5.0
pyparsing==3.1.2
pyrsistent==0.19.3
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
pytz==2024.1
pywin32==306
pywin32-ctypes==0.2.3
pywinpty==2.0.10
pyzmq==26.0.3
qtconsole==5.4.4
QtPy==2.4.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.2
Send2Trash==1.8.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.4.1
SQLAlchemy==2.0.36
sshtunnel==0.4.0
sympy==1.10.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
terminado==0.17.1
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
tinycss2==1.2.1
tornado==6.2
tqdm==4.66.4
traitlets==5.9.0
typing_extensions==4.7.1
urllib3==2.0.7
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.6.1
Werkzeug==2.2.3
widgetsnbextension==4.0.11
wincertstore==0.2
wrapt==1.16.0
zipp==3.15.0
